+++ src/app/(auth)/login/_types/login.interface.ts
@@ -1,12 +1,6 @@
-import type { SubmitHandler } from 'react-hook-form';
-
 export interface ILogin {
   email: string;
   password: string;
-}
-
-export interface ILoginSubmit {
-  onSubmit: SubmitHandler<ILogin>;
 }
 
 export interface ILoginResponse {
